Added openstreetmap data example, fixed broken example link in README.md, added a image overlay of NYC (copyright-free) that's not animated, and added some advice
