+++ packages/examples/README.md
@@ -1,0 +1,4 @@
+# Examples folder
+This is the folder with examples. You can either run the expo-app folder in packages/expo-app or packages/react-native-app.
+
+Do note that there are features that currently don't work! Check out this issue to see what's missing (and what you can fix): https://github.com/maplibre/maplibre-react-native/issues/12
