+++ packages/examples/src/assets/empty-json-file.json
@@ -1,0 +1,1 @@
+{}
