+++ packages/examples/src/examples/Map/ShowMapLocalStyle.tsx
@@ -6,7 +6,7 @@ import sheet from "../../styles/sheet";
 import Page from "../common/Page";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const style = JSON.stringify(require("../../assets/map-styleURL-style.json"));
+const style = JSON.stringify(require("../../assets/openstreet-map-style.json"));
 
 const ShowMap: FC<any> = (props) => {
   useEffect(() => {
